+++ src/main/java/com/project/bookstore/web/user/UsersController.java
@@ -1,24 +1,22 @@
 package com.project.bookstore.web.user;
 
-import com.project.bookstore.config.ApiResponse;
 import com.project.bookstore.config.ApiResponse2;
+import com.project.bookstore.service.users.AddrService;
+import com.project.bookstore.service.users.CardService;
 import com.project.bookstore.service.users.UsersService;
-import com.project.bookstore.web.user.dto.cardDto.CardInsertDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-
-import java.util.List;
 
 @RequiredArgsConstructor
 @Controller
 public class UsersController {
 
     private final UsersService usersService;
+    private final AddrService addrService;
+    private final CardService cardService;
 
     @GetMapping("/")
     public String index() {
@@ -47,7 +45,7 @@ public class UsersController {
 //        model.addAttribute("userInfo", usersService.findAll(users_id));
         ApiResponse2 result = null;
         try{
-            result = new ApiResponse2(true, "성공", usersService.findById(users_ID), usersService.findCard(users_ID), usersService.findAddr(users_ID));
+            result = new ApiResponse2(true, "성공", usersService.findById(users_ID), cardService.findCard(users_ID), addrService.findAddr(users_ID));
             return ResponseEntity.ok().body(result);
         } catch (Exception e) {
             e.printStackTrace();
